review_mvp/requirements.txt
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.3
pytest==8.3.5
